Refactored Code to eventually support Combining DataMaps.
Attempting to add functionality that Edmund’s little toy has over to my
Scala library right now, and as such, I’m trying to figure out a way to
combine two maps into one map. However, I found that I needed a status
map for this new map, and this combiner is in a different class, and so
I’m needing to refactor a bit of the code that was originally in the
ComputeStep class into their own class.