Added Error Tracking, Validated Default Multi-Threading
Added capabilities for an error tracker, with a list of exceptions
which will be compared with the Errors. (If there are actually errors
that are in there. If not, then the default code will still occur)
Finished my multi-threading Akka code and verified that the default
configurations of that code worked. I still have yet to try adding a
config file and have yet to test Error Tracking, though…