Started a Brief Implementation.
Right now it’s only transformer, but a few other operators will also
have a stepActor that will be implemented in a global ActorSystem. I
have yet to figure out how to do the micro-batch implementation,
though. That will either be in the StepActor code, or it will be in the
process code for the Transformers.