Added some way to Call Steps via Actors
Slowly making this more Actor based. Just need to link steps together,
and it should be good!
Created a StepActor and a ComposedPipeActor, which handle the different
steps. Both are state-based via become, and are able to add more steps
that they need to pass along stuff when they need it. (For stuff like
Junctions) The ComposedPipeActor and just the Composed step took up
most of the work here, I think, and will probably take up most of the
work with the linking.
Unfortunately, through the process of attempting to link stuff
together, as the actors need another ActorSystem, we may have
(temporarily?) lost the ability to use Objects in our pipeline, as we
need to pass in the ActorSystem. However, this may not be too big of a
deal since we may have needed to do this anyway to pass in the Config
files…