Linked together and works, but lost a few things.
Every step is now linked together, and is running as it did within the
single threaded version!
There are a few things that we have lost through my implementation,
though.
   Config file benefits probably lost: Since the Akka actors that
controlled each step were moved, its workers were also moved, so the
actor config stuff is now broken. This can be fixed kind of easily, but
it may be easier just to have it so the config files for every step get
merged into a larger config file at the beginning of the program or
something. I’ll think about how to fix this.
   Debug messages are missing. This should be an easy fix?