Changing config file passing to Any
Most of the time, the config files are just going to be strings that
are then parsed by Akka, and so it’s recommended to use a String value.
However, there are also points where you will want to load a config
object in by itself, so that’s still supported.