Removed Deprecated Computes
Computes have been deprecated, so I rerouted into the mappers. That
being said, there still is quite a bit of copy-paste code. If I can get
to it, I would like to add single computes to the actual computations.
(Or at least to Reducer. I had to rewrite half of the run code for
reducer to get it to work with a single job…)