Added Crashing Code, worked on SolrMap
Added code when the workers crash for whatever reason.
Made it so you can add jobs a list at a time, and made that the default
for both the BigActorUnaryPlatform and the BigActorBinaryPlatform
Added some put functions into the SolrMap. Now for the get functions,
which I’m really nervous about getting right…