Created a FileSystemIterator
Created a FileSystemIterator based off of my FileSystemMap. Trying to
figure out how to conceptualize the FileSystemDataMap, and I’m a bit
lost trying to do so… :| While I’m working on conceptualizing the
MultiDataMap, I’m also working on the letter that I have to write for
the DLA so I feel like I’m making more progress on something.