Fixed stupid bug on SolrDataMap; Changed canRemove to hasSchema
Turns out, there was a bug in the saveTheOmitted function that
duplicated the list as many times as there were objects in the list.
Besides that, I fixed another bug that I forgot to catch when making
sure JsArrays could have JsStrings in them. (How that slipped by me,
I’m not entirely sure…) The FileSystemIterator is also now connected to
the FileSystemDataMap. Next step: Talk to Edmund and see what’s needed
to do next. (Maybe writing more examples for SolrDataMap?)