Wrote some functions for the SolrBackend
I wrote some functions for the SolrBackend to allow for nested maps.
However, I feel as though there may be a better way of doing this, or
at least I feel as though there is some sort of refactoring that I may
need to do here, so I’m not quite adding it into the public functions
just yet.