Misc. Bug Fixes and trying to debug FeatureExtractor
Got rid of the update debug command in SolrBackend.
Stabilized BigActors so sending two Wake commands at the same time does
not crash it.
Tried to figure out why FeatureExtractor is so picky to no avail… :(
Added the protobuf for the next parts, but still have to integrate that
into the project.